Rename getColor => _getColor
