+++ packages/theming/src/utils/_getColor.spec.ts
@@ -5,14 +5,14 @@
  * found at http://www.apache.org/licenses/LICENSE-2.0.
  */
 
-import { getColor } from './getColor';
+import { getColor } from './_getColor';
 import PALETTE from '../elements/palette';
 import DEFAULT_THEME from '../elements/theme';
 import { darken, lighten, rgba } from 'polished';
 
 const DEFAULT_SHADE = 600;
 
-describe('getColor', () => {
+describe('_getColor', () => {
   describe('by hue', () => {
     it('gets the hue specified by string', () => {
       const color = getColor('red');
