+++ packages/theming/src/utils/_getColor.ts
